requirements.txt
# Core services
selenium
openai
crawl4ai

pandas
openpyxl
networkx
scikit-learn

pdfminer.six
PyMuPDF

# Dev Supplements
pip
python-dotenv

# Database
pymongo